pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["sentinel"]

[tool.setuptools.package-data]
sentinel = ["py.typed"]  # If you want to support type hints

[project]
name = "sentinel"
version = "0.1.0"
description = "A Python SDK for Sentinel"
authors = [
    { name = "Enrtropy Labs", email = "devs@entropy-labs.ai" }
]
license = { file = "LICENSE" }
readme = "README.md"
dependencies = [
    "requests>=2.28.0",
    "openai>=1.0.0",
    "anthropic>=0.3.0",
    "litellm>=0.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0"
]
